feat : LottoInputView class 구현 및 domain과 view의 package 분리.
Input view class의 각 구현요소와 유효성검사를 구현하고, domain과 view의 패키지를 분리하여 각 class file을 위치시켰음.
