+++ src/main/java/lotto/domain/LottoNos.java
@@ -1,10 +1,8 @@
-package lotto;
+package lotto.domain;
 
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 import java.util.HashSet;
-import java.util.stream.Collector;
 import java.util.stream.Collectors;
 
 public class LottoNos {
