+++ src/test/java/lotto/domain/LottoNosTest.java
@@ -1,5 +1,6 @@
-package lotto;
+package lotto.domain;
 
+import lotto.domain.LottoNos;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
